pyproject.toml
[tool.poetry]
name = "receipttosheet"
version = "0.1.0"
description = ""
authors = ["Julia <jtl.lab.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = ">=1.22,<=1.24.3"
pandas = "^2.0.3"
pdf2image = "^1.16.3"
pytesseract = "^0.3.10"
tesseract = "^0.1.3"
opencv-python = "^4.8.0.74"
pillow = "^10.0.0"
matplotlib = "^3.7.2"
scikit-learn = "^1.3.0"
tensorflow = "^2.13.0"
gradio = "^3.38.0"
keras = "^2.13.1"
neptune-tensorflow-keras = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
